Pengecekan tombol yang ditekan untuk update dan delete, master ikan paging, penjual bibit diganti supplier

+++ application/controllers/Masterblok.php
@@ -126,21 +126,25 @@ class Masterblok extends CI_Controller {
 
         $this->data['id'] = $this->input->post('tid');
         $this->data['name'] = $this->input->post('tname');
-        if ($this->form_validation->run() != FALSE)
-        {
-            if($this->Blok->cek_kembar($this->data['name'], $this->data['id']) == false){
-                $this->data['msg'] = "<div id='err_msg' class='alert alert-danger sldown' style='display:none;'>Nama blok kembar</div>";
-            } else {
-                $result = $this->Blok->update($this->data['name'], $this->data['id'], $_SESSION['id']);
-                if($result == 1){
-                    redirect('Masterblok');
-                }else{
-                    $this->data['msg'] = "<div id='err_msg' class='alert alert-danger sldown' style='display:none;'>Update Gagal</div>";
+        if($this->input->post('write') == "write"){
+            if ($this->form_validation->run() != FALSE)
+            {
+                if($this->Blok->cek_kembar($this->data['name'], $this->data['id']) == false){
+                    $this->data['msg'] = "<div id='err_msg' class='alert alert-danger sldown' style='display:none;'>Nama blok kembar</div>";
+                } else {
+                    $result = $this->Blok->update($this->data['name'], $this->data['id'], $_SESSION['id']);
+                    if($result == 1){
+                        redirect('Masterblok');
+                    }else{
+                        $this->data['msg'] = "<div id='err_msg' class='alert alert-danger sldown' style='display:none;'>Update Gagal</div>";
+                    }
                 }
             }
+            $this->data["state"] = "update";
+            $this->load->view('masterblok_form', $this->data);
+        } else {
+            redirect('Masterblok');
         }
-        $this->data["state"] = "update";
-        $this->load->view('masterblok_form', $this->data);
     }
 
 
@@ -148,14 +152,18 @@ class Masterblok extends CI_Controller {
         $this->check_role();
         $this->initialization();
         $this->data['id'] = $this->input->post('tid');
-        $result = $this->Blok->delete($this->data['id'], $_SESSION['id']);
-        if($result == 1){
+        if($this->input->post('delete') == "delete") {
+            $result = $this->Blok->delete($this->data['id'], $_SESSION['id']);
+            if($result == 1){
+                redirect('Masterblok');
+            }else{
+                $this->data['msg'] = "<div id='err_msg' class='alert alert-danger sldown' style='display:none;'>Hapus Data Gagal</div>";
+            }
+            $this->data["state"] = "delete";
+            $this->load->view('masterblok_form', $this->data);
+        } else {
             redirect('Masterblok');
-        }else{
-            $this->data['msg'] = "<div id='err_msg' class='alert alert-danger sldown' style='display:none;'>Hapus Data Gagal</div>";
         }
-        $this->data["state"] = "delete";
-        $this->load->view('masterblok_form', $this->data);
     }
